move options menu into deprecated dir
