+++ packages/react-core/src/components/index.ts
@@ -51,7 +51,7 @@ export * from './MultipleFileUpload';
 export * from './Nav';
 export * from './NotificationBadge';
 export * from './NotificationDrawer';
-export * from './OptionsMenu';
+export * from '../deprecated/components/OptionsMenu';
 export * from './OverflowMenu';
 export * from './Page';
 export * from './Pagination';
