+++ packages/react-core/src/deprecated/components/OptionsMenu/OptionsMenuItem.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { css } from '@patternfly/react-styles';
 import styles from '@patternfly/react-styles/css/components/OptionsMenu/options-menu';
-import { DropdownItem } from '../Dropdown';
+import { DropdownItem } from '../../../components/Dropdown';
 import CheckIcon from '@patternfly/react-icons/dist/esm/icons/check-icon';
 
 export interface OptionsMenuItemProps
