+++ packages/react-core/src/deprecated/components/OptionsMenu/OptionsMenuSeparator.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Divider, DividerProps } from '../Divider';
+import { Divider, DividerProps } from '../../../components/Divider';
 
 export const OptionsMenuSeparator: React.FunctionComponent<DividerProps> = ({
   component = 'li',
